fix: final-review nits from small-scalar-width whole-branch review
Three tiny follow-ups from the phase's final review, now that bool/char
are 1 byte in aggregates:

- internal/mactest/uiprobe/probe_ui.c: TableRow.done and FormRec.active
  were still int32_t under RT_FT_BOOL descriptors, stale against the
  1-byte runtime ABI. File is frozen/out of the build path (pending 5f
  deletion), so this is cosmetic consistency only.
- docs/ROADMAP.md: forward-pointer the 5e narrative's "full 4-byte slot"
  sentence to the since-narrowed 1-byte width.
- docs/ROADMAP.md: file the pre-existing cross-lane string(n) record-field
  alignment divergence (cg68k 2-byte vs cprint 1-byte, found during Task 4)
  as a small open item, to resolve or document before Ch13 freezes wording.

+++ internal/mactest/uiprobe/probe_ui.c
@@ -365,7 +365,7 @@ static const rt_ui_window_desc kPopupProbeWindow = {
  * PopupProbe's own precedent. */
 
 typedef struct { unsigned char name[21]; /* 1 len + 20 data bytes (strCap 20) */
-                 int32_t qty; int32_t done; } TableRow;
+                 int32_t qty; uint8_t done; } TableRow;
 
 static rt_list *gTableRows = NULL;
 
@@ -480,7 +480,7 @@ static const rt_ui_window_desc kTableProbeWindow = {
  * RTUI_EV_ACCEPTED/CANCELLED only ever hand it FormProbe's own inst. */
 
 typedef struct { unsigned char name[21]; /* 1 len + 20 data bytes (strCap 20) */
-                 int32_t qty; int32_t active; int32_t color;
+                 int32_t qty; uint8_t active; int32_t color;
                  int32_t price; /* RT_FT_FIXED, 16.16 -- Fix round 1 coverage
                                     for rt_ui_parse_fixed's whole-number bug */
 } FormRec;
